+++ bench/bench.ml
@@ -45,9 +45,8 @@ module Context = struct
 
     let v () = random_string key_size
 
-    let hash = Hashtbl.hash
-
-    let hash_size = 30
+    let hash x =
+      Hashtbl.hash x |> Int64.of_int |> fun h -> Int64.shift_left h (64 - 30)
 
     let encode s = s