+++ test/fuzz/fan/main.ml
@@ -1,8 +1,6 @@
 open Crowbar
 module Fan = Index.Private.Fan
 
-let hash_size = 30
-
 let entry_size = 56
 
 let entry_sizeL = Int64.of_int entry_size
@@ -11,12 +9,12 @@ let int_bound = 100_000_000
 
 let bounded_int = map [ int ] (fun i -> abs i mod int_bound)
 
-let hash = map [ bytes ] Hashtbl.hash
+let hash = int64
 
-let hash_list = map [ list hash ] (List.sort compare)
+let hash_list = map [ list hash ] (List.sort Int64.unsigned_compare)
 
 let empty_fan_with_size =
-  map [ bounded_int ] (fun n -> (Fan.v ~hash_size ~entry_size n, n))
+  map [ bounded_int ] (fun n -> (Fan.v ~entry_size n, n))
 
 let empty_fan = map [ empty_fan_with_size ] fst
 
@@ -29,7 +27,7 @@ let update_list =
 
 let fan_with_updates =
   map [ empty_fan; update_list ] (fun fan l ->
-      List.iter (fun (hash, off) -> Fan.update fan hash off) l;
+      List.iter (fun (hash, offset) -> Fan.update fan ~hash ~offset) l;
       Fan.finalize fan;
       (fan, l))
 
@@ -42,7 +40,7 @@ let check_export_size fan =
 
 let check_export fan =
   let exported = Fan.export fan in
-  let imported = Fan.import ~hash_size exported in
+  let imported = Fan.import exported in
   check_eq ~eq:Fan.equal imported fan
 
 let check_updates (fan, updates) =
@@ -50,11 +48,8 @@ let check_updates (fan, updates) =
     (fun (hash, off) ->
       let low, high = Fan.search fan hash in
       if off < low || high < off then
-        (* Use Crowbar.failf on next release *)
-        fail
-          (Printf.sprintf
-             "hash %d was added at off %Ld, but got low=%Ld, high=%Ld" hash off
-             low high))
+        Crowbar.failf "hash %Ld was added at off %Ld, but got low=%Ld, high=%Ld"
+          hash off low high)
     updates
 
 let check_fan_size (fan, size) =